turnos

+++ src/main/java/com/informatorio/app/service/EventServiceImpl.java
@@ -5,13 +5,10 @@ import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.crossstore.ChangeSetPersister.NotFoundException;
 import org.springframework.stereotype.Service;
-import org.springframework.web.client.HttpClientErrorException.BadRequest;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.informatorio.app.dto.request.EventDto;
 import com.informatorio.app.dto.response.EventResponseDto;
-import com.informatorio.app.entity.Appointment;
 import com.informatorio.app.entity.Event;
 import com.informatorio.app.entity.Organization;
 import com.informatorio.app.exception.AlreadyExistException;
@@ -19,13 +16,13 @@ import com.informatorio.app.exception.BadRequestException;
 import com.informatorio.app.exception.InvalidPasswordException;
 import com.informatorio.app.repository.IEventDao;
 import com.informatorio.app.repository.IOrgDao;
+import com.informatorio.app.utils.CheckDate;
 import com.informatorio.app.wrapper.EventWrapper;
-import com.informatorio.app.wrapper.OrganizationWrapper;
 
 @Service
 public class EventServiceImpl implements IEventService {
 
-	private static final Logger log = LoggerFactory.getLogger(EventServiceImpl.class);
+	
 
 	@Autowired
 	IOrgDao orgDao;
@@ -37,7 +34,11 @@ public class EventServiceImpl implements IEventService {
 			throws AlreadyExistException, InvalidPasswordException, BadRequestException {
 
 		Organization org = orgDao.findById(eventDto.getOrganizationId()).orElse(new Organization());
-
+		
+		if (eventDto.getIsUnique() && eventDto.getEventHour() == null && eventDto.getEventDate() == null) {
+			throw new BadRequestException("A unique event must have an hour and a date");
+		}
+		
 		if (!eventDao.findByIsActiveAndByNameAndOrg(eventDto.getOrganizationId(), eventDto.getName()).isEmpty()) {
 			throw new AlreadyExistException("This event already exist");
 		}
@@ -46,7 +47,7 @@ public class EventServiceImpl implements IEventService {
 			Event event = EventWrapper.dtoToEntity(eventDto);
 			event.setOrganization(org);
 			event.setEventDate(eventDto.getEventDate());
-			EventDto dto =  EventWrapper.entityToDto(eventDao.save(event));
+			EventDto dto = EventWrapper.entityToDto(eventDao.save(event));
 			return EventWrapper.dtoToResponse(dto);
 		} else {
 
@@ -59,12 +60,16 @@ public class EventServiceImpl implements IEventService {
 	public List<Event> findByAll() {
 
 		List<Event> event = eventDao.findAll();
-
+		event.stream().forEach(e -> e.setIsActive(CheckDate.check(e.getEventDate())));
 
 		return event;
 	}
 
-	public void delete(Long id) {
+	public void delete(Long id,Organization request) throws NotFoundException, InvalidPasswordException {
+		if (eventDao.findById(id).isEmpty())throw new NotFoundException();
+		Organization org = eventDao.findById(id).orElse(new Event()).getOrganization();
+		if (!org.getPassword().equals(request.getPassword()))throw new InvalidPasswordException("Invalid password");
+		
 
 		eventDao.deleteById(id);
